A workaround about compatibility between company-dabbrev and company-yasnippet-autoparens.
When typing `=`, no corresponding `(=)` appears in the completion list in scheme-mode or racket-mode, but OK in emacs-lisp-mode.

The reason is that when typing `=`, `company--multi-backend-adapter` will be called (with command=prefix backends=(company-capf company-dabbrev company-yasnippet-autoparens company-yasnippet) args=nil), and the following code:

```
((or `prefix `ignore-case `no-cache `require-match)
       (let (value)
         (cl-dolist (backend backends)
           (when (setq value (company--force-sync
                              backend (cons command args) backend))
             (cl-return value)))))
```
in the emacs-lisp-mode, `company--force-sync` be called once and return `=`:

```
company--force-sync, backend = company-capf
company--force-sync, value is =
company--multi-backend-adapter value is =
```

but in scheme-mode, `company--force-sync`  be called twice, and return `nil` and `""`:

```
company--force-sync, backend = company-capf
company--force-sync, value is nil
company--force-sync, backend = company-dabbrev
company--force-sync, value is
company--multi-backend-adapter value is
```

That because

1. emacs-lisp-mode hooks `completion-at-point-functions` which handle `=` (N.B. `company-capf` is a bridge between completion and company), but no `completion-at-point-functions` hook in scheme-mode, so it return `nil`.

2. company-dabbrev return `""` rather than `nil` for prefix `=`, so company-mode thought the correct prefix was returned.

A workaround is add `=` to `company-dabbrev-char-regexp`. Also add `<` and `>`.