Single-semicolon comment lines NOT indented in scheme-mode!
N.B. Lisp language (e.g. Elisp) has a standard convention that a single ; is indented to the right (see https://www.gnu.org/software/emacs/manual/html_node/elisp/Comment-Tips.html), but it is not the case in Scheme. However, scheme-mode use `lisp-indent-line` for `indent-line-function`, that is not ideal.